+++ client/src/components/v2/OverlayMessage.stories.tsx
@@ -1,0 +1,1 @@
+export default {};
